requirements.txt
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain>=0.1.0
langchain-huggingface>=0.0.2
pandas>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.12.0
numpy>=1.24.0
pydantic>=2.0.0
selenium~=4.26.1